requirements.txt
faiss-cpu
flask==3.0.3
matplotlib
langchain==0.2.10
langchain-community==0.2.10
langchain-core==0.2.43
langchain-ollama==0.1.0
langchain-text-splitters==0.2.4
pandas
python-dotenv==1.0.1
sentence-transformers==3.3.1
sentencepiece==0.2.0